// src/generated/prisma/enums.ts
export const StaffRole = {
  OWNER: 'OWNER',
  MANAGER: 'MANAGER',
  CASHIER: 'CASHIER',
  KITCHEN: 'KITCHEN',
  WAITER: 'WAITER'
} as const

export type StaffRole = (typeof StaffRole)[keyof typeof StaffRole]